export built

+++ Modules/Export/Http/Controllers/Admin/ExportController.php
@@ -9,6 +9,7 @@ use Modules\Export\Http\Requests\ExportCreateRequest;
 use Modules\Export\Http\Requests\ExportUpdateRequest;
 use Modules\Export\Http\Resources\ExportResource;
 use Modules\Export\Models\Export;
+use Modules\Export\Repositories\ExportRepository;
 use Modules\Export\Services\ExportService;
 use Modules\Export\Services\ExportStorage;
 
@@ -16,13 +17,14 @@ class ExportController extends Controller
 {
     public function __construct(
         protected ExportStorage $exportStorage,
+        protected ExportRepository $exportRepository,
         protected ExportService $exportService
-    ) {
-        $this->authorizeResource(Export::class, 'export');
-    }
+    ) {}
 
     public function store(ExportCreateRequest $request)
     {
+        $this->authorize('create', Export::class);
+
         $exportDto = ExportDto::fromFormRequest($request);
 
         if (!$exportDto->assigned_by_id) {
@@ -34,22 +36,32 @@ class ExportController extends Controller
         return new ExportResource($export);
     }
 
-    public function update(Export $export, ExportUpdateRequest $request)
+    public function update(int $export, ExportUpdateRequest $request)
     {
+        $export = $this->exportRepository->find($export);
+
+        $this->authorize('update', $export);
+
         $export = $this->exportStorage->update($export, ExportDto::fromFormRequest($request));
 
         return new ExportResource($export);
     }
 
-    public function destroy(Export $export)
+    public function destroy(int $export)
     {
+        $export = $this->exportRepository->find($export);
+
+        $this->authorize('delete', $export);
+
         $this->exportStorage->delete($export);
 
         return response()->noContent();
     }
 
-    public function export(Export $export)
+    public function export(int $export)
     {
+        $export = $this->exportRepository->find($export);
+
         $this->authorize('export', $export);
 
         $generator = $this->exportService->getGenerator($export);
